#1020: rename combinators file
